Create a comparison story between the old and new static question
