+++ enterprise/frontend/src/embedding-sdk/components/public/StaticQuestion/StaticQuestion.stories.tsx
@@ -6,8 +6,10 @@ import {
   questionIdArgType,
   questionIds,
 } from "embedding-sdk/test/storybook-id-args";
+import { Stack } from "metabase/ui";
 
 import { StaticQuestion } from "./StaticQuestion";
+import { StaticQuestionOLD } from "./StaticQuestionOLD";
 
 const QUESTION_ID = (window as any).QUESTION_ID || questionIds.numberId;
 
@@ -31,6 +33,31 @@ const Template: StoryFn<StaticQuestionComponentProps> = args => {
 
 export const Default = {
   render: Template,
+
+  args: {
+    questionId: QUESTION_ID,
+    isSaveEnabled: true,
+    saveToCollection: undefined,
+  },
+};
+
+export const Old = {
+  render: args => <StaticQuestionOLD {...args}></StaticQuestionOLD>,
+
+  args: {
+    questionId: QUESTION_ID,
+    isSaveEnabled: true,
+    saveToCollection: undefined,
+  },
+};
+
+export const Comparison = {
+  render: args => (
+    <Stack>
+      <StaticQuestion {...args}></StaticQuestion>
+      <StaticQuestionOLD {...args}></StaticQuestionOLD>
+    </Stack>
+  ),
 
   args: {
     questionId: QUESTION_ID,
